+++ resolwe/flow/models/data.py
@@ -341,6 +341,10 @@ class Data(BaseModel):
 
             entity.data.add(self)
 
+            # Inherite collections from entity.
+            for collection in entity.collections.all():
+                collection.data.add(self)
+
     def save(self, render_name=False, *args, **kwargs):  # pylint: disable=keyword-arg-before-vararg
         """Save the data model."""
         # Generate the descriptor if one is not already set.